requirements.txt
pandas==2.1.0
numpy==1.26.4
requests>=2.32.0
redis
prettytable
pymysql
chardet

pyecharts==2.0.8
dbutils
apscheduler
tenacity
flask
flask-login
gevent
tornado
sqlalchemy
lark-oapi

mytt
matplotlib
tqdm
ipywidgets
pinyin
pyarmor
setuptools

futu-api
ccxt
baostock
polygon-api-client
alpaca-py
ib_insync>=0.9.86
akshare>=1.16.0
tqsdk
openai
playwright
pyyaml

pyfolio-reloaded
dtaidistance